another reek down.
Instead of creating an instance of NoIRC class to do the null object pattern,
we use a the Module instance NoIRC module. This gets rid of reek
UtilityMethod, as it did not rely on instance state, but duh, it's null object
it doesn't have instance state. Using the module itself get rid of the
unneeded instantiation.